requirements.txt
discord.py==2.2.3
praw==7.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
SQLAlchemy==2.0.12
